+++ gw2buildutil/api/crawl.py
@@ -84,8 +84,8 @@ def crawl (client=gw2client.Client(),
         return
 
     crawler = Crawler(client, storage, entity_types)
-    if storage.schema_version() != gw2client.SCHEMA_VERSION or full_recrawl:
+    if storage.schema_version() != client.schema_version or full_recrawl:
         storage.clear_raw()
-        storage.store_schema_version(gw2client.SCHEMA_VERSION)
+        storage.store_schema_version(client.schema_version)
     storage.clear() # required if entity definition changes
     crawler.crawl_all(entity_types)